requirements.txt
certifi==2018.1.18
configparser==3.5.0
future==0.16.0
python-telegram-bot==9.0.0
